reorganizing tests and adding integration test
